flask/microservice1/requirements.txt
Flask
paho-mqtt
requests
aiocoap
aiohttp
CoAPthon3